v1.3.4
added NRopt output for estimated yld from the ffopt and act scenario. Act refers to what the yld,pro,NR would be if the actual as-applied experiment was applied in the simulation year.

made an outline for an ObsOP class

added to the SimOP savePlots method to make maps of each of the yld, pro, NR from each of the mgmt strategies

changed color schemes of maps to red yellow green for yld, pro, N applied, NR, veg indxs to match convention. prec = blue, gdd = yellow orange red