v1.7.14
big update that fixes warnings from using dbSendQuery using the Postgres() driver. This means that no wranings are made that say cancelling previous query etc. to clean up the code. Also did some other stuff but that is the biggest so far.

oh it also adds a step in the SimClass where the nr sim does one rep using the economic conditions from the simulation year to find the ssopt etc. and actual net-return using the economic conditions from that year. The other figures use the mean from all economic conditions to come up with one. This is good for comparison if the sim year is the same as the most recent year on record or an observed year for example